working solution for part 1 of the 7th of december

+++ 7December/solution.h
@@ -1,5 +1,6 @@
 #ifndef SOLUTION7_H
 #define SOLUTION7_H
-int solution7();
+#include <bits/stdc++.h>
+uint64_t solution7();
 int second_solution7();
 #endif
